Report position information on semantic errors
Now that creating a node is external to the grammar, and the node
function is passed the parser object, this adds a reference to the
parser to an AST node, and adds the current position of the node.

This allows us to get ~approximately correct position information
for semantic errors. I haven't really dug into this yet to figure
out how far off the positions are, but there's probably room for
issues around the starting and ending positions of a rule/node,
and it's possible I'm looking at the wrong node's position when
reporting the error as well. I will continue to iterate on this
and fix issues as I find them.