Move node()-using json5 grammar to grammars/json5_cst.g.
At the moment I'm not sure how to make a single grammar co-exist
for node()-using things and non-node()-using things. node() by itself
works fine, but other functions that expect non-node values
(like `dict()`) will get confused if they are handed nodes instead
of raw values.

It's possible we need to invoke functions two different ways
depending on whether the function is built-in or an extern (and
then you'd override the builtin non-node-expecting functions with
extern versions that did expect nodes), but I'm not actually sure
if that would even solve the problem.