Clean up CST grammar a bit.
We don't need nodes for 'member_list', 'element_list', so just use
regular lists for those. We also don't really need nodes for 'members',
but using one makes the structure of objects a bit simpler (otherwise
we'd need to have object['c'] be a list of lists or a flattened list
of pairs.