Add 'pfunc' extern type in addition 'func' type.
This introduces a distinction between two different
types of extern functions: ones which require a parser object
as an argument, and ones which don't. In the json5 grammar,
`node` does, `dict` doesn't. By making this distinction, we
can now reuse the same grammar for both a tokenized syntax tree
parser a value-only parser by just offering different implementations
of node() and dict().

As part of this change, I remove the separate grammars/json5_cst.g
grammar; it is no longer needed.