Merge in a bunch of work intended to make full parse trees possible.
For many parsing applications, you actually need position and
token information along with just generic data being passed back
from the semantic actions; in some cases you need to be able to
reproduce the input exactly, e.g., when you want to pretty-print
a document or edit it programmatically.

This series of changes (from the 'cst' branch) makes that possible.
It adds new built-in `node` and `pos` functions; `pos` can return
the current position in the grammar, and the `node` function is
passed the parser as an argument and can query the parser for
additional state.

It also adds the ability to 'tokenize' a grammar, meaning that
when requested, all tokens, whitespace, and comments are saved
in a list in the parser along with their positional information.
As one example, using the `node()` function, you can then query
the tokens and associate them with nodes; this can be used to
build a full parse tree (aka concrete syntax tree). The tokenization
code so far as only been implemented in the python code generator,
and not in the interpreter or the other two generators, but it
should be easy to port it over.

As part of this work, I removed the "_ws" versions of the
sample grammars, as there no longer seemed to be a point to
keeping them around. I also added a more fully worked-out
example of some JSON5 code in //examples/json5, containing
the start of what could be a round-trip editor. Currently, it
doesn't actually support doing any editing, but it does build
out an AST and a CST using the same grammar, compares that the
two match, and checks that the CST matches the input text.

There are a bunch of miscellaneous fixes and cleanup of other
things as well.

I have not yet added any tests for any of this work, so the
coverage is poor and things are probably kinda fragile. I
wanted to get what I had merged in before I proceeded on to that.